requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
attrs==24.2.0
    # via -r requirements/base.txt
certifi==2024.8.30
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.4.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.7
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==1.8.1
    # via -r requirements/test.in
coverage[toml]==7.6.7
    # via pytest-cov
cryptography==43.0.3
    # via
    #   -r requirements/base.txt
    #   pyjwt
ddt==1.7.2
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-model-utils
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-model-utils==5.0.0
    # via -r requirements/base.txt
django-waffle==4.1.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.15.2
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
dnspython==2.7.0
    # via
    #   -r requirements/base.txt
    #   pymongo
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-django-utils==7.0.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   edx-rest-api-client
edx-drf-extensions==10.5.0
    # via -r requirements/base.txt
edx-opaque-keys==2.11.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-rest-api-client==6.0.0
    # via -r requirements/base.txt
freezegun==1.5.1
    # via -r requirements/test.in
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via
    #   -r requirements/base.txt
    #   code-annotations
markupsafe==3.0.2
    # via
    #   -r requirements/base.txt
    #   jinja2
newrelic==10.2.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
packaging==24.2
    # via pytest
pbr==6.1.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.5.0
    # via pytest
psutil==6.1.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pyjwt[crypto]==2.9.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
    #   edx-rest-api-client
pymongo==4.10.1
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==8.3.3
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.9.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via freezegun
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.2
    # via
    #   code-annotations
    #   responses
requests==2.32.3
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   responses
responses==0.25.3
    # via -r requirements/test.in
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
six==1.16.0
    # via python-dateutil
sqlparse==0.5.2
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.3.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.12.2
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
urllib3==2.2.3
    # via
    #   -r requirements/base.txt
    #   requests
    #   responses
